design improvements, move some code
